+++ app/src/components/message-form/index.ts
@@ -1,0 +1,1 @@
+export { MessageForm } from './message-form';
